Split HIR capability validation (#1302)
Move FFI validation, capability checks, network and process allowlist validation, and capability-focused HIR tests from hir.rs into hir/capabilities.rs while preserving compiler.hir behavior. Update the compiler source monolith reporter and ratchet ceilings for the reduced HIR facade line count.

Refs #1254.
